requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.20.33
botocore==1.23.33
celery==5.2.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.7
django-cors==0.1
django-cors-headers==3.8.0
django-filter==2.4.0
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-writable-nested==0.6.3
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.2.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
mysqlclient==2.1.0
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.2
prompt-toolkit==3.0.24
protobuf==3.19.3
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==4.1.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
simplejwt==2.0.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
